New function for item removal
Implemented an alternative function for removing an item from the Inventory. Instead of a name, it takes a number representing the index slot to take item(s) from.

By changing the UseMenu to use this new function, used/sold items now decrement the chosen stack instead of always taking from the last stack